OpenSoftBackend/gc_backend/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
asgiref==3.8.1
asyncio==3.4.3
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
bitsandbytes==0.45.4
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.0.0
datasets==3.5.0
dill==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
eval_type_backport==0.2.2
Faker==16.0.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mistralai==1.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.70.0
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==10.2.0
propcache==0.3.1
psutil==7.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
shap==0.47.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.0
trl==0.16.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zope.interface==7.2
